Módulo 7: Código final Ktor Client Ejercicio

+++ modulo7_client_ej1/src/main/java/com/devexperto/modulo7_client_ej1/ui/activities/AddClienteActivity.kt
@@ -75,7 +75,7 @@ class AddClienteActivity : AppCompatActivity() {
 
     private fun eliminarCliente() {
         lifecycleScope.launch {
-            val respose = cliente?.let { service.deleteCliente(it.id.toInt()) }
+            val respose = cliente?.let { service.deleteCliente(it.id.toLong()) }
             Toast.makeText(this@AddClienteActivity, respose?.message ?: ":", Toast.LENGTH_SHORT).show()
             finish()
         }
